% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.data.frame.ontology_index}
\alias{as.data.frame.ontology_index}
\title{Case \code{ontology_index} object to \code{data.frame}}
\usage{
\method{as.data.frame}{ontology_index}(x, ...)
}
\arguments{
\item{x}{\code{ontology_index} object.}

\item{...}{Arguments pass to \code{as.data.frame}.}
}
\value{
\code{data.frame} containing information about an ontology, rows corresponding to terms and columns to properties.
}
\description{
Case \code{ontology_index} object to \code{data.frame}
}
