% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check}
\alias{check}
\title{Perform simple consistency checks on \code{ontology_index} object}
\usage{
check(ontology, stop_if_invalid = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{stop_if_invalid}{Logical value determining whether the function should call \code{stop} and print an error message upon finding that the given \code{ontology_index} is in valid.}
}
\description{
Perform simple consistency checks on \code{ontology_index} object
}

