% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{exclude_branch}
\alias{exclude_branch}
\title{Exclude terms descending from particular term from a character vector of terms}
\usage{
exclude_branch(ontology, branch_root, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{branch_root}{Term whose descendants should be excluded}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Exclude terms descending from particular term from a character vector of terms
}

