% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{exclude_branches}
\alias{exclude_branches}
\title{Exclude terms descending from particular term from a character vector of terms}
\usage{
exclude_branches(ontology, branch_roots, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{branch_roots}{Character vector of term IDs of branch roots.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Exclude terms descending from particular term from a character vector of terms
}

