% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{exclude_descendants}
\alias{exclude_descendants}
\title{Exclude terms descending from any in a given set of root terms}
\usage{
exclude_descendants(ontology, roots, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{roots}{Character vector of IDs for terms with respect to which descendants are to be defined.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Exclude from set \code{terms}, any terms that are either in, or descend from one of, the set \code{roots}.
}
\seealso{
\code{\link{intersection_with_descendants}}, \code{\link{prune_descendants}}
}

