% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Get set of terms containing all ancestors of terms in a given set}
\usage{
get_ancestors(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of all terms which are an ancestor of at least one term in \code{terms}, including the terms themselves
}
\description{
Get set of terms containing all ancestors of terms in a given set
}
\examples{
data(hpo)
get_ancestors(hpo, c("HP:0001873", "HP:0011877"))
}
\seealso{
\code{link{get_descendants}}
}

