% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_descendants}
\alias{get_descendants}
\title{Get set of all descendants of single term}
\usage{
get_descendants(ontology, ancestor, remove_ancestor = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{ancestor}{Character vector of length 1 - the ID of the term whose descendants you wish to retrieve}

\item{remove_ancestor}{Boolean indicating whether to remove the given ancestor or not}
}
\value{
Character vector of terms
}
\description{
Get set of all descendants of single term
}
\examples{
data(hpo)
get_descendants(hpo, ancestor=c("HP:0001873"))
}
\seealso{
\code{link{get_ancestors}}
}

