% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_mpo_to_hpo}
\alias{get_mpo_to_hpo}
\title{Get MPO to HPO list}
\usage{
get_mpo_to_hpo(ontology, cross_species_file)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{cross_species_file}{cross species .obo file, available via http://human-phenotype-ontology.org}
}
\value{
List of HPO terms per MPO term
}
\description{
Get MPO to HPO list
}

