% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ontology}
\alias{get_ontology}
\title{Get R-Object representation of ontology from obo file}
\usage{
get_ontology(file, qualifier = "HP")
}
\arguments{
\item{file}{File path of obo file}

\item{qualifier}{Character vector - "HP" for HPO, "MP" for MPO, etc.}
}
\value{
R-Object (list) representing ontology
}
\description{
Get R-Object representation of ontology from obo file
}

