% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_relation_names}
\alias{get_relation_names}
\title{Get names of relations used in OBO file}
\usage{
get_relation_names(file)
}
\arguments{
\item{file}{File path of OBO formatted file.}
}
\description{
Get names of relations used in OBO file
}
\seealso{
\code{\link{get_ontology}}
}

