% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_term_property}
\alias{get_term_ancestors}
\alias{get_term_children}
\alias{get_term_names}
\alias{get_term_parents}
\alias{get_term_property}
\title{Get property of individual ontological term}
\usage{
get_term_property(ontology, property_name, term)

get_term_names(ontology, terms)

get_term_children(ontology, term, as_names = FALSE)

get_term_parents(ontology, term, as_names = FALSE)

get_term_ancestors(ontology, term, as_names = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{property_name}{Name of property.}

\item{term}{Character value of term ID.}

\item{terms}{Character vector of ontological terms.}

\item{as_names}{Logical value determining whether to return character vector of names or IDs (default).}
}
\description{
Get property of individual ontological term
}

