% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{intersection_with_branches}
\alias{intersection_with_branches}
\title{Intersect set of terms with branches of ontology}
\usage{
intersection_with_branches(ontology, branch_roots, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{branch_roots}{Character vector of roots of branches you wish to intersect with}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Intersect set of terms with branches of ontology
}
\examples{
data(hpo)
intersection_with_branches(hpo, "HP:0001872", c("HP:0001873", "HP:0011877"))
}

