% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{intersection_with_descendants}
\alias{intersection_with_descendants}
\title{Intersect a set of terms with the descendants of a given set of roots}
\usage{
intersection_with_descendants(ontology, roots, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{roots}{Character vector of IDs for terms with respect to which descendants are to be defined.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Intersect a set of terms with the descendants of a given set of roots
}
\examples{
data(hpo)
intersection_with_descendants(hpo, c("HP:0001872", "HP:0000707"), c("HP:0001873", "HP:0011877"))
}
\seealso{
\code{\link{exclude_descendants}}, \code{\link{prune_descendants}}
}

