% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{minimal_set}
\alias{minimal_set}
\title{Remove redundant/implied terms from a set of terms}
\usage{
minimal_set(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Remove redundant/implied terms from a set of terms
}
\examples{
data(hpo)
minimal_set(hpo, c("HP:0001873", "HP:0001872"))
}

