% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpo_to_hpo}
\alias{mpo_to_hpo}
\title{Object containing data for mapping between MPO and HPO}
\format{List of HPO terms per MPO term}
\description{
List containing cross-species ontology (MPO to HPO) information - character vectors of HPO terms indexed by associated MPO term IDs
}

