\name{ontologyIndex}
\docType{package}
\title{
	Functions for Reading Ontologies into R	
}
\description{
Functions for reading ontologies into R as lists and handling sets of ontological terms.
}
\details{
\tabular{ll}{
Package: \tab ontologyIndex\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-11\cr
License: \tab GPL (>= 2)\cr
}
The key function is \code{\link{get_ontology}}, which creates the index, and object of class \code{ontology_index}.
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
\references{
	`The Human Phenotype Ontology project: linking molecular biology and disease through phenotype data', Nucl. Acids Res. (1 January 2014) 42 (D1): D966-D974 doi:10.1093/nar/gkt1026
}
