% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{OWL_properties}
\alias{OWL_IDs}
\alias{OWL_labels}
\alias{OWL_obsolete}
\alias{OWL_properties}
\alias{OWL_strings_from_nodes}
\title{Get vectors of term properties from set of OWL nodes}
\usage{
OWL_strings_from_nodes(xpath)

OWL_IDs(nodes)

OWL_labels(nodes)

OWL_obsolete(nodes)
}
\arguments{
\item{xpath}{XPath expression.}

\item{nodes}{\code{xml_nodeset} object where nodes correspond to ontological terms.}
}
\description{
Functions which return vectors of properties for OWL formatted ontological terms represented as \code{xml_nodeset}s. \code{\link{OWL_strings_from_nodes}} accepts an XPath expression (a \code{character} vector of length 1) and returns the string value of the XPath expression evaluated for each node (i.e. for each term in the ontology). One can use \code{\link{OWL_list_attributes_per_node}} to create functions for extracting lists of term attributes for each node of encoded in the OWL file.
}
\seealso{
\code{\link{OWL_list_attributes_per_node}}
}

