% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.ontology_index}
\alias{print.ontology_index}
\title{Print \code{ontology_index} object.}
\usage{
\method{print}{ontology_index}(x, ...)
}
\arguments{
\item{x}{\code{ontology_index} object.}

\item{...}{Unused parameters.}
}
\value{
Prints a summary
}
\description{
Print \code{ontology_index} object.
}

