% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{prune_branch}
\alias{prune_branch}
\title{Prune all terms descending from given term down to that term and ensure no degeneracy}
\usage{
prune_branch(ontology, prune_to_point, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{prune_to_point}{Character value equal to term ID which can be included, but whose descendants should be excluded}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Warning! Prunes down whole branch, even if terms have parents which don't descend from prune point.
}

