% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{prune_branches}
\alias{prune_branches}
\title{Prune all terms descending from given term down to that term and ensure no degeneracy}
\usage{
prune_branches(ontology, branch_roots, terms)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{branch_roots}{Character vector of term IDs of branch roots.}

\item{terms}{Character vector of ontological terms.}
}
\value{
Character vector of terms
}
\description{
Warning! Prunes down whole branch, even if terms have parents which don't descend from prune point.
}

