% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{get_descendants}
\alias{get_descendants}
\title{Get set of terms containing all descendants of terms in a given set}
\usage{
get_descendants(ontology, roots, exclude_roots = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{roots}{Character vector of IDs for terms with respect to which descendants are to be defined.}

\item{exclude_roots}{Boolean determining whether to remove the given \code{roots} terms from the result.}
}
\value{
Character vector of terms
}
\description{
Get set of terms containing all descendants of terms in a given set
}
\examples{
data(hpo)
get_descendants(hpo, roots="HP:0001873")
}
\seealso{
\code{link{get_ancestors}}
}

