% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ontology}
\alias{get_ontology}
\title{Get R-Object representation of ontology from obo file}
\usage{
get_ontology(file, include_descriptions = FALSE)
}
\arguments{
\item{file}{File path of obo file}

\item{include_descriptions}{Logical value determining whether to parse the term descriptions and include them in the \code{ontology_index} object.}
}
\value{
R-Object (list) representing ontology
}
\description{
Get R-Object representation of ontology from obo file
}

