% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ontology}
\alias{get_ontology}
\title{Read ontology from obo file into R}
\usage{
get_ontology(file, ...)
}
\arguments{
\item{file}{File path of obo file}

\item{...}{Additional arguments to pass onto \code{\link{get_OWL}} or \code{\link{get_OBO}}.}
}
\value{
R-Object (list) representing ontology
}
\description{
Reads an ontology from a file in either an OBO or OWL format into R as an \code{ontology_index} object.
}
\seealso{
\code{\link{get_OWL}} \code{\link{get_OBO}}
}

