% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_term_property}
\alias{get_term_property}
\title{Get property of individual ontological term}
\usage{
get_term_property(ontology, property_name, term, as_names = FALSE)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{property_name}{Name of property.}

\item{term}{Character value of term ID.}

\item{as_names}{Logical value determining whether to return character vector of names (defaults to \code{FALSE}).}
}
\description{
Get property of individual ontological term
}

